requirements.txt
aiohttp>=3.8
sqlalchemy[asyncio]>=1.4
asyncpg>=0.27
alembic>=1.10
pydantic>=1.10
ccxt>=2.0
python-dotenv
pytest
httpx
fastapi
uvicorn[standard]
pywebpush>=1.9.0
bcrypt>=3.2.0,<4
redis>=4.5
numpy
aiogram==3.0.0b7
tradingview_websocket
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
python-multipart
qrcode[pil]>=7.4.2
click>=8.0.0